Create 16TH April 2023 (3) --w3 html basic
 --- different types of heading tags ---<h1>,<h2>,<h3>,<h4>,<h5>,<h6>
 ---insert link using <a href = "www.zyx.com">link</a> tag
 ---insert image using <img src ="Mule.jpg or file path " alt="mule image not loading"  width="140"  height ="140"> 
Images are not technically inserted into a web page; images are linked to web pages. The <img> tag creates a holding space for the referenced image.

EXTRA INFO:- BUT USEFUL
#How to View HTML Source
View HTML Source Code:
Right-click in an HTML page and select "View Page Source" (in Chrome) or "View Source" (in Edge), or similar in other browsers. This will open a window containing the HTML source code of the page.

Inspect an HTML Element:
Right-click on an element (or a blank area), and choose "Inspect" or "Inspect Element" to see what elements are made up of (you will see both the HTML and the CSS). You can also edit the HTML or CSS on-the-fly in the Elements or Styles panel that opens.